feat: replaced mypy with ty, added tflint and minor refactors
