--- .vscode/settings.json
@@ -1,8 +1,0 @@
-{
-    "python.testing.pytestArgs": [
-        "--no-cov",
-        "tests"
-    ],
-    "python.testing.unittestEnabled": false,
-    "python.testing.pytestEnabled": true
-}
