--- mypy.ini
@@ -1,2 +1,0 @@
-[mypy]
-plugins = pydantic.mypy
