+++ pyproject.toml
@@ -7,7 +7,6 @@ requires-python = ">=3.14"
 dependencies = [
     "argon2-cffi>=25.1.0",
     "boto3>=1.42.45",
-    "boto3-stubs[essential]>=1.42.45",
     "fastapi>=0.128.6",
     "httpx>=0.28.1",
     "mangum>=0.21.0",
@@ -25,7 +24,6 @@ dev = [
     "aws-lambda-powertools[validation]>=3.24.0",
     "bandit>=1.9.3",
     "moto>=5.1.21",
-    "mypy>=1.19.1",
     "pytest>=9.0.2",
     "pytest-cov>=7.0.0",
     "pytest-env>=1.2.0",
@@ -34,4 +32,5 @@ dev = [
     "respx>=0.22.0",
     "ruff>=0.15.0",
     "time-machine>=3.2.0",
+    "ty>=0.0.15",
 ]
