+++ pytest.ini
@@ -1,6 +1,5 @@
 [pytest]
 addopts = --cache-clear --cov-branch --cov-report term --cov=app/ -n 4 -r f
-asyncio_mode = strict
 env =
     APP_NAME=personal-backend-service
     AWS_ACCESS_KEY_ID=access_key_id
