+++ tflint.hcl
@@ -1,0 +1,5 @@
+plugin "aws" {
+  enabled = true
+  version = "0.45.0"
+  source  = "github.com/terraform-linters/tflint-ruleset-aws"
+}
